add support for MCU clock > 8MHz
Original code generated a 1 cycle I2C clock pulse. This works fine on 8MHz MCU but not for 16MHz CPU.
Added code that when compiled for MCU with a clock > 8MHz a 2 cycle cbi instruction will be used intead of a single cycle out instruciton extending the I2C clock pulse.